// src/API.ts
export interface OrdersDto {
  address?: string | null;
  discountAmount?: number | null;
  id?: number | null;
  memberId?: number;
  memberName?: string | null;
  message?: string | null;
  orderDate?: string;
  paymentMethod?: string | null;
  phone?: string | null;
  status?: string | null;
  totalAmount?: number;
}